refactor(lenses): composable for html lens
